+++ src/main/kotlin/de/andrena/tools/altn8th/domain/relatedFiles/find/strategies/prefix/FindRelatedFilesByPrefixStrategy.kt
@@ -16,7 +16,7 @@ internal class FindRelatedFilesByPrefixStrategy : FindRelatedFilesStrategy {
         val relationType = baseNameToPrefixSettings.mapNotNull { (basename, originHop) ->
             val relatedFileHop = settings.prefixes.firstOrNull { relatedFileHop ->
                 areNotIdentical(origin, file)
-                    && areRelated(basename, file.nameWithoutFileExtension(), relatedFileHop)
+                    && areRelated(basename, file.nameWithoutFileExtension(), relatedFileHop, settings)
             }
             relatedFileHop?.let { PrefixRelationType(originHop, it) }
         }.firstOrNull()
@@ -25,10 +25,22 @@ internal class FindRelatedFilesByPrefixStrategy : FindRelatedFilesStrategy {
 
     private fun areNotIdentical(origin: File, relatedFile: File): Boolean = origin != relatedFile
 
-    private fun areRelated(basename: String, relatedFile: String, relatedFileHop: PrefixSetting) =
-        areRelatedByGivenPattern(basename, relatedFile, relatedFileHop.pattern)
-            || areRelatedByGivenPattern(relatedFile, basename, relatedFileHop.pattern)
+    private fun areRelated(
+        basename: String,
+        relatedFile: String,
+        relatedFileHop: PrefixSetting,
+        settings: SettingsState
+    ) =
+        areRelatedByGivenPattern(basename, relatedFile, relatedFileHop.pattern, settings)
+            || areRelatedByGivenPattern(relatedFile, basename, relatedFileHop.pattern, settings)
 
-    private fun areRelatedByGivenPattern(first: String, second: String, prefixPattern: String): Boolean =
-        first.matches(Regex("^(${prefixPattern})${second}$", RegexOption.IGNORE_CASE))
+    private fun areRelatedByGivenPattern(
+        first: String,
+        second: String,
+        prefixPattern: String,
+        settings: SettingsState
+    ): Boolean {
+        val regexOptions = if (settings.caseInsensitiveMatching) setOf(RegexOption.IGNORE_CASE) else emptySet()
+        return first.matches(Regex("^(${prefixPattern})${second}$", regexOptions))
+    }
 }